fixed paging bug in server-side

+++ api/DTOs/Vehicle/VehicleQueryDTO.cs
@@ -7,6 +7,6 @@ namespace api.DTOs.Vehicle
         public string SortBy { get; set; } = string.Empty;
         public bool IsSortAscending { get; set; }
         public int Page { get; set; }
-        public int PageSize { get; set; }
+        public byte PageSize { get; set; }
     }
 }
